phpstan fixes

+++ src/Module/Util/Waveform.php
@@ -337,7 +337,7 @@ class Waveform
         // fill background of image
         // transparent background specified
         imagesavealpha($img, true);
-        $transparentColor = imagecolorallocatealpha($img, 0, 0, 0, 127);
+        $transparentColor = (int)imagecolorallocatealpha($img, 0, 0, 0, 127);
         imagefill($img, 0, 0, $transparentColor);
         while (!feof($handle) && $data_point < $data_size) {
             if ($data_point++ % $detail == 0) {
@@ -386,7 +386,7 @@ class Waveform
                         $height - $value, // x2
                         (int)($data_point / $detail), // y2: same as y1, but from the bottom of the image
                         $height - ($height - $value),
-                        imagecolorallocate($img, (int)$red, (int)$green, (int)$blue)
+                        (int)imagecolorallocate($img, (int)$red, (int)$green, (int)$blue)
                     );
                 }
             } else {
